requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
pydantic
python-dotenv
python-jose
databases